python_backend/pyproject.toml
[tool.poetry]
name = "obsidian-agent"
version = "0.1.0"
description = "Python backend for Obsidian Agent plugin - RAG, vector search, and advanced indexing"
authors = ["B0LK13"]
license = "MIT"
readme = "README.md"
packages = [{include = "obsidian_agent", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
typer = {extras = ["all"], version = "^0.9.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
chromadb = "^0.4.24"
sentence-transformers = "^2.2.2"
tiktoken = "^0.5.2"
sqlalchemy = "^2.0.23"
alembic = "^1.13.1"
fastapi = {version = "^0.108.0", optional = true}
uvicorn = {version = "^0.25.0", optional = true}
watchdog = "^3.0.0"
pymupdf = {version = "^1.23.0", optional = true}
pytesseract = {version = "^0.3.10", optional = true}
pillow = {version = "^10.0.0", optional = true}
python-frontmatter = "^1.0.1"
markdown = "^3.5.1"
httpx = "^0.26.0"
cryptography = {version = "^41.0.0", optional = true}
openai-whisper = {version = "^20231117", optional = true}
rich = "^13.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.1"
black = "^23.12.1"
isort = "^5.13.2"
mypy = "^1.7.1"
ruff = "^0.1.9"

[tool.poetry.extras]
api = ["fastapi", "uvicorn"]
ocr = ["pymupdf", "pytesseract", "pillow"]
encryption = ["cryptography"]
voice = ["openai-whisper"]

[tool.poetry.scripts]
obsidian-agent = "obsidian_agent.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=obsidian_agent --cov-report=term-missing"

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]
